Move DuplicateInjection test to TypeAnalyzerTest
