--- VContainer/Assets/UnitTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7ea3f561be6ad4dadb6557c0c60a662e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
