--- VContainer/Assets/UnitTests/DuplicateInjectionTest.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: df5c9e81a1fd499289c23dd51861364a
-timeCreated: 1719851024
