--- VContainer/Assets/UnitTests/VContainer.UnitTests.asmdef
@@ -1,38 +1,0 @@
-{
-    "name": "VContainer.UnitTests",
-    "rootNamespace": "",
-    "references": [
-        "UnityEngine.TestRunner",
-        "UnityEditor.TestRunner",
-        "VContainer",
-        "Unity.Collections",
-        "Unity.Entities",
-        "UniTask"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": true,
-    "precompiledReferences": [
-        "nunit.framework.dll"
-    ],
-    "autoReferenced": false,
-    "defineConstraints": [
-        "UNITY_INCLUDE_TESTS"
-    ],
-    "versionDefines": [
-        {
-            "name": "com.cysharp.unitask",
-            "expression": "",
-            "define": "VCONTAINER_UNITASK_INTEGRATION"
-        },
-        {
-            "name": "com.unity.entities",
-            "expression": "",
-            "define": "VCONTAINER_ECS_INTEGRATION"
-        }
-    ],
-    "noEngineReferences": false
-}
