--- VContainer/Assets/UnitTests/VContainer.UnitTests.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e0776455dbb0c4601a0243fcdc6e8a10
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
