+++ VContainer/Assets/VContainer/Runtime/AssemblyInfo.cs
@@ -1,7 +1,6 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("VContainer.Tests")]
-[assembly: InternalsVisibleTo("VContainer.UnitTests")]
 [assembly: InternalsVisibleTo("VContainer.StandaloneTests")]
 [assembly: InternalsVisibleTo("VContainer.Editor")]
 [assembly: InternalsVisibleTo("Unity.VContainer.CodeGen")]
